refactor: split party server.go into handler endpoint files
Split the 1359-line server.go into focused files by operation group:
- server.go (149 lines): struct, constructor, With* methods, errors, interfaces
- party_handlers.go (433 lines): RegisterParty, RetrieveParty, ListParties, UpdateParty, ControlParty
- reference_handlers.go (107 lines): UpdateReference, RetrieveReference
- association_handlers.go (225 lines): RegisterAssociations, UpdateAssociations, RetrieveAssociations, ListParticipants, GetStructuringData
- demographics_handlers.go (222 lines): ExchangeDemographics, UpdateDemographics, RetrieveDemographics, UpdateBankRelations, RetrieveBankRelations
- proto_converters.go (289 lines): all proto/domain conversion functions
